use decimal data type for population data
to ensure arithmetic in later models is accurate

+++ products/cdbg/models/product/cdbg_boroughs.sql
@@ -9,7 +9,7 @@ eligibility_calculation AS (
         round(total_floor_area::numeric)::integer AS total_floor_area,
         round(residential_floor_area::numeric)::integer AS residential_floor_area,
         round(residential_floor_area_percentage::numeric, 2) AS residential_floor_area_percentage,
-        low_mod_income_population::bigint,
+        low_mod_income_population::integer,
         round(low_mod_income_population_percentage::numeric, 2) AS low_mod_income_population_percentage,
         low_mod_income_population_percentage >= 51 AND residential_floor_area_percentage >= 50 AS eligibility_flag
     FROM boros
